nlp-service/requirements.txt
flask==3.0.0
spacy==3.7.2
python-dateutil==2.8.2
nltk==3.8.1
gunicorn==21.2.0
